compilers: Add submodule and clean duplicate pkgs
All of the packages present in the 'compilers' submodule are deleted
from this repository.

CI now will only build OSX versions of those packages, but will use
recipes from the 'compilers' submodule. This is temporary – until the
'litex-conda-compilers' CI scripts will be prepared to build packages
on macOS and Windows.